以前に作業していた内容を全プッシュ 以降こまめに管理する事

+++ eshacu-backend/src/main/java/dev/eshacu/api/dto/common/PageResponse.java
@@ -1,4 +1,5 @@
 package dev.eshacu.api.dto.common;
 
 import java.util.List;
-public record PageResponse<T>(List<T> items, int page, int pageSize, int totalRecords) { }
+
+public record PageResponse<T>(List<T> items, int page, int pageSize, long totalRecords) { }
